COM-1038 fix from pr comment
